Install terraform when the applier starts up
Use the tf-install package to grab the desired version of terraform when
the applier starts up. This means you don't need to re-release
terraform-applier for every new version.

It also means you don't need to preinstall terraform when running the
applier outside of the docker image.

Also includes the option to use a specific terraform binary that already
exists on the system.